refactor widgets/screens to be in sub packages of framework
