+++ common/src/main/java/org/vivecraft/client/gui/framework/VROptionEntry.java
@@ -1,6 +1,7 @@
 package org.vivecraft.client.gui.framework;
 
 import net.minecraft.world.phys.Vec2;
+import org.vivecraft.client.gui.framework.widgets.GuiVROption;
 import org.vivecraft.client_vr.settings.VRSettings;
 
 import java.util.function.BiFunction;
