+++ common/src/main/java/org/vivecraft/client/gui/framework/widgets/GuiVROption.java
@@ -1,4 +1,4 @@
-package org.vivecraft.client.gui.framework;
+package org.vivecraft.client.gui.framework.widgets;
 
 import net.minecraft.client.gui.layouts.LayoutElement;
 import org.vivecraft.client_vr.settings.VRSettings;
