+++ common/src/main/java/org/vivecraft/client/gui/framework/widgets/SettingsList.java
@@ -1,4 +1,4 @@
-package org.vivecraft.client.gui.widgets;
+package org.vivecraft.client.gui.framework.widgets;
 
 import com.google.common.collect.ImmutableList;
 import net.minecraft.client.Minecraft;
@@ -14,7 +14,6 @@ import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.resources.language.I18n;
 import net.minecraft.network.chat.Component;
 import net.minecraft.util.Mth;
-import org.vivecraft.client.gui.framework.GuiVROptionSlider;
 import org.vivecraft.client_vr.ClientDataHolderVR;
 import org.vivecraft.client_vr.settings.VRSettings;
 import org.vivecraft.server.config.ConfigBuilder;
