+++ common/src/main/java/org/vivecraft/client/gui/settings/GuiPlayerModelSettings.java
@@ -2,7 +2,7 @@ package org.vivecraft.client.gui.settings;
 
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.Screen;
-import org.vivecraft.client.gui.framework.GuiVROptionsBase;
+import org.vivecraft.client.gui.framework.screens.GuiVROptionsBase;
 import org.vivecraft.client.gui.framework.VROptionEntry;
 import org.vivecraft.client.gui.screens.FBTCalibrationScreen;
 import org.vivecraft.client_vr.ClientDataHolderVR;
