+++ common/src/main/java/org/vivecraft/client/gui/settings/GuiRadialConfiguration.java
@@ -7,8 +7,8 @@ import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.resources.language.I18n;
 import net.minecraft.network.chat.Component;
 import org.apache.commons.lang3.ArrayUtils;
-import org.vivecraft.client.gui.framework.GuiVROptionButton;
-import org.vivecraft.client.gui.framework.GuiVROptionsBase;
+import org.vivecraft.client.gui.framework.widgets.GuiVROptionButton;
+import org.vivecraft.client.gui.framework.screens.GuiVROptionsBase;
 import org.vivecraft.client.gui.framework.VROptionLayout;
 import org.vivecraft.client_vr.gui.GuiRadial;
 import org.vivecraft.client_vr.settings.VRSettings;
