+++ common/src/main/java/org/vivecraft/client/gui/settings/GuiWeaponCollisionSettings.java
@@ -1,7 +1,7 @@
 package org.vivecraft.client.gui.settings;
 
 import net.minecraft.client.gui.screens.Screen;
-import org.vivecraft.client.gui.framework.GuiVROptionsBase;
+import org.vivecraft.client.gui.framework.screens.GuiVROptionsBase;
 import org.vivecraft.client_vr.settings.VRSettings;
 
 public class GuiWeaponCollisionSettings extends GuiVROptionsBase {
